HDFont, HDFontLayer, DMD window pos & other stuff
dmd changes:
- added config options dmd_offset_x and dmd_offset_y that specify the
window position in your OS
- using smoothscale on the dot filter to avoid dot distortion
Frame (in dmd.c):
- added new blend modes: mult, min, max and in addition to blacksrc
there's now magentasrc and greensrc
- added a scale operation for Frames (it's CPU intensive, don't do it
often)
- scale can also scale all frames if it's any type of frame collection
hdFont.py:
- Added a 'line width' option so that rendered text can have a different
border color (be creative, and use magenta, green or black for
compositing tricks with groupedLayers)
Layers.py:
- fixed a bug in PanningLayer when bounce was false, layers wouldn't
keep moving when they hit the edge; now layers can leave the screen
- PanningLayers can specify how long they are hold the frame between
movements (allows for extra slow panning)
- NEW: HDTextLayer - a text layer that supports the border width option
so you can specify an interior color, exterior color, and background --
font colors are specified at the time the layer is created
- NEW: AnimatedHDTextLayer: builds a layer for you that will show your
text and play an animation as the interior of the text.  This is mostly
a helper and may not be ready for most people's use cases just yet...
markup.py:
- Assuming you are using HDFonts (why aren't you) you can specify the
plain_font and the bold_font with an interior, border color and border
width.  Then using the old markup gives you nicer looking colored and
edged text
- There's also an option to pass in a collection of FrameLayers (or
AnimationLayers) and if you specify {key} in the markup, the image found
in the first frame with the matching key will be inserted into the
generated markup.
Game.py:
- shows number of DMD updates (e.g., FPS) --thanks Josh!
- got rid of the failure callback that was added for Visual Pinball
support -- I added something else in the VP bridge instead that does the
same job without mucking up this.
tool: dmdconvert:
- fixed a bug that prevented converting old style, gray-scale dmd files
to new full-color dmd files (which is actually pretty cool).

General: Removed a bunch of random print statements that seem harmless
but can cause the Visual Pinball bridge to crash (if you care about that
sort of thing)